ThermalCamera/Desktop/requirements.txt
numpy==1.19.4
Pillow==8.0.1
pyzmq==20.0.0
opencv_contrib_python==4.4.0.46
Logic==0.2.3
